revert: refactor
