+++ src/components/project-tools-table.tsx
@@ -12,7 +12,7 @@ import {
 import { Dependency } from '@/constants/types'
 import { useQueryState } from 'nuqs'
 import { useMemo } from 'react'
-import { Cell } from '@/components/d11/DataCells'
+import { Cell } from '@/components/DataCells'
 import { EndOfLifeCell } from './EndOfLifeCell'
 import { LatestVersionCell } from './LatestVersionCell'
